### 2024-05-30 #6807 @editorialbot
[PRE REVIEW]: DWBuilder: A code to generate ferroelectric/ferroelastic domain walls and multi-material atomic interface structures
**Submitting author:** @mzkhalid039 (<a href="http://orcid.org/0000-0002-7866-3870">Muhammad Zeeshan Khalid</a>)
**Repository:** https://github.com/mzkhalid039/DWBuilder
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
