### 2024-05-30 #6807 @kyleniemeyer
@mzkhalid039 due to the relatively small nature of your software (in terms of code), the JOSS editorial board is going to take a look to see if this meets our requirements for substantial scholarly effort. We should get back to you with a decision within a week or two.
