+++ README.md
@@ -1,6 +1,5 @@
 ![GitHub release (latest by date)](https://img.shields.io/github/v/release/mzkhalid039/DWBuilder)
 ![Python Version](https://img.shields.io/badge/python-3.11-blue)
-![GitHub stars](https://img.shields.io/github/stars/mzkhalid039/DWBuilder)
 
 # DWBuilder: Domain Wall Builder
 Ferroelectric/ferroelastic domain_wall_builder creates domain walls and interface structure for the perovskites and ceramics. The code reads the structure in the vasp format, and then creates domain walls with user-defined sizes and angles.
