### 2024-06-17 #6895 @mzkhalid039
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/9532756257?check_suite_focus=true). Problem with ORCID (https://orcid.org/0000-0002-7866-3870) for M Khalid. ORCID looks to be the wrong length.

I have fixed this problem. I hope it will be able to generate pdf now.
