### 2024-06-25 07841d1
set numpy dependency <2.0
