### 2024-07-06 e658c50
Delete DWBuilder.egg-info directory
