+++ DWBuilder.egg-info/requires.txt
@@ -1,3 +1,0 @@
-ase
-numpy
-matplotlib
