+++ DWBuilder.egg-info/entry_points.txt
@@ -1,3 +1,0 @@
-[console_scripts]
-dbuilder = scripts.dbuilder:main
-dwbuilder = scripts.dwbuilder:main
