+++ DWBuilder.egg-info/SOURCES.txt
@@ -1,10 +1,0 @@
-setup.py
-DWBuilder.egg-info/PKG-INFO
-DWBuilder.egg-info/SOURCES.txt
-DWBuilder.egg-info/dependency_links.txt
-DWBuilder.egg-info/entry_points.txt
-DWBuilder.egg-info/requires.txt
-DWBuilder.egg-info/top_level.txt
-scripts/__init__.py
-scripts/dbuilder.py
-scripts/dwbuilder.py
