+++ scripts/dwbuilder.py
@@ -35,814 +35,431 @@
 """
 
 import os
-import sys
-import ase
 import ase.io
 from ase.build import cut, rotate, stack
-import numpy as np
-from colorama import init, Fore, Style
+from ase.neighborlist import NeighborList
 from numpy.linalg import norm
-from colorama import Fore, Style
 from pymatgen.core.structure import Structure
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
-
-# Find the path of the current directory
-current_path = os.getcwd()
-
-filename = input("Enter the input file name (with extension): ")
-D1 = ase.io.read(filename)
-
-struct = Structure.from_file(filename)
-sym = SpacegroupAnalyzer(struct)
-data = sym.get_symmetry_dataset()
-
-print("Space group number: {}".format(data["number"]))
-print("International symbol: {}".format(data["international"]))
-print(f"Lattice type: {sym.get_lattice_type()}")
-
-# Conditionally select the system based on international symbol
-if data["international"] == "R3c":
-    system_selection = 1
-elif data["international"] == "R3m":
-    system_selection = 2
-elif data["international"] == "P4mm":
-    system_selection = 3
-elif data["international"] == "P6_3cm":
-    system_selection = 4
-elif data["international"] == "Pnma":
-    system_selection = 5
-elif data["international"] == "Pmc2_1":
-    system_selection = 6
-
-else:
-    print("Warning: The international symbol does not match any pre-defined systems.")
-    print("Please manually select the system.")
-    system_selection = 7
-
-# Prompt the user to select the system if not automatically selected
-if system_selection == 7:
-    print("Select the system:")
-    print("1 - R3c")
-    print("2 - R3m")
-    print("3 - P4mm")
-    print("4 - P6_3cm")
-    print("5 - Pnma")
-    print("6 - Pmc2_1")
-
-    print("7 - Manual Orientation relationships")
-
-
-    system_selection = int(input())
-
-# BFO system 
-# Reference: Wang Y, Nelson C, Melville A, Winchester B, Shang S, Liu ZK, Schlom DG, Pan X, Chen LQ. BiFeO 3 domain wall energies and structures: a combined experimental and density functional theory+ U study. Physical review letters. 2013 Jun 24;110(26):267601.
-
-if system_selection == 1:
-
-    # Cut and rotate the structure to obtain a pseudo-cubic structure
-    pseudo_cubic = cut(D1, a=[1.0, 1.0, -1], b=[-1, 1, 1], c=[1, -1, 1])
-    rotate(pseudo_cubic, pseudo_cubic.cell[0], (0,1,0), pseudo_cubic.cell[1], (1,0,0))
-
-    # Write the pseudo-cubic structure to a VASP file
-    pseudo_cubic.write('Pseudo_cubic_BFO.vasp', sort=True, vasp5=True)
-
-    # Read the pseudo-cubic structure
-    BFO = ase.io.read('Pseudo_cubic_BFO.vasp')
-
-    # Prompt the user to select the domain wall angle
-    print("Select the domain wall angle:")
-    print("1 - 109 degrees")
-    print("2 - 71 degrees")
-    print("3 - 180 degrees")
-    print("4 - All angles")
-    angle_selection = int(input())
-
-    # Prompt the user to enter the domain wall size
[…]
