+++ setup.py
@@ -1,10 +1,9 @@
 from setuptools import setup, find_packages
-from os.path import abspath, dirname
 
 setup(
     name='dwbuilder',
-    version='1.1.0',
-    authors = 'M.Z.Khalid, S.M.Selbach',
+    version='2.0.0',
+    author='M.Z.Khalid, S.M.Selbach',
     author_email='zeeshan.khalid039@gmail.com',
     description='A package for developing domain wall structures for atomistic calculations',
     license='MIT',
@@ -16,14 +15,16 @@ setup(
         'Operating System :: OS Independent',
     ],
     python_requires='>=3.6',
-    install_requires=['ase','numpy<2','matplotlib', 'pymatgen', 'colorama'],
+    install_requires=[
+        'ase',
+        'numpy<2',
+        'matplotlib',
+        'pymatgen',
+        'colorama'
+    ],
     entry_points={
         'console_scripts': [
-            'dwbuilder = scripts.dwbuilder:main',
-            'dbuilder = scripts.dbuilder:main',
-            'hibuilder = scripts.hibuilder:main',
-            'polarization = scripts.polarization:main',
+            'dwbuilder = main:main',
         ]
     }
 )
-
