### 2024-07-18 18c1650
Delete examples directory
