+++ examples/Pnma_FDW/LOGFILE.txt
@@ -1,17 +1,0 @@
-Space group number: 62
-International symbol: Pnma
-Lattice type: orthorhombic
-System selection: 5
-Domain wall angle: DomainWallAngle.FDW
-Domain wall size: 4 unit cells
-Supercell size: a=2, b=1, c=1
-Cutoff distance: 0.5 angstroms
-Stacked slab saved to FDW_stacked.vasp
-Domain slab 1 saved to FDW_domain1.vasp
-Domain slab 2 saved to FDW_domain2.vasp
-Lattice strain for FDW DW:
-strain along a (%): 0.00
-strain along b (%): 0.00
-strain along c (%): 0.00
-Supercell saved to FDW_supercell.vasp
-Domain wall structures and supercells created successfully!
