### 2024-07-18 6890ee7
Delete scripts/main.py
