+++ main.py
@@ -17,7 +17,9 @@ def main():
         '1': 'dwbuilder.py',
         '2': 'dbuilder.py',
         '3': 'hibuilder.py',
-        '4': 'slab.py'
+        '4': 'slab.py',
+        '5': 'polarization.py'
+
     }
 
     print("Select a script to run:")
