+++ README.md
@@ -90,6 +90,14 @@ Amm2 (user-defined)
 In this example, I have reproduced an orientation relationship reported in the following article [CMS](https://www.sciencedirect.com/science/article/pii/S0927025621000446). 
 
 ```
+(py311) muhammzk:Jupyter zeeshan$ dwbuilder
+Select a script to run:
+1: dwbuilder.py
+2: dbuilder.py
+3: hibuilder.py
+4: slab.py
+5: polarization.py
+Enter the number of the script to run: 3
 Enter the bulk phase 1file name (with extension): Fe_unitcell.vasp
 Enter the bulk phase 2 file name (with extension): Fe2Al5.vasp
 Bulk phase 1 Space group number: 229
