+++ scripts/hibuilder.py
@@ -25,114 +25,160 @@
 """
 
 import os
-import sys
+import shutil
 import ase
 import ase.io
 from ase.build import cut, rotate, stack
 import numpy as np
 from colorama import init, Fore, Style
 from numpy.linalg import norm
-from colorama import Fore, Style
 from pymatgen.core.structure import Structure
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
+from ase.neighborlist import NeighborList
 
 
-filename1 = input("Enter the bulk phase 1file name (with extension): ")
-filename2 = input("Enter the bulk phase 2 file name (with extension): ")
+def read_structure(file_name):
+    return ase.io.read(file_name)
 
 
-# Read the input file
-D1 = ase.io.read(filename1)
-D2 = ase.io.read(filename2)
+def get_symmetry_info(file_name, log):
+    struct = Structure.from_file(file_name)
+    sym = SpacegroupAnalyzer(struct)
+    data = sym.get_symmetry_dataset()
+    log.append(f"Space group number: {data['number']}")
+    log.append(f"International symbol: {data['international']}")
+    log.append(f"Lattice type: {sym.get_lattice_type()}")
+    print(f"Space group number: {data['number']}")
+    print(f"International symbol: {data['international']}")
+    print(f"Lattice type: {sym.get_lattice_type()}")
 
 
-struct = Structure.from_file(filename1)
-sym = SpacegroupAnalyzer(struct)
-data = sym.get_symmetry_dataset()
-
-print("Bulk phase 1 Space group number: {}".format(data["number"]))
-print("Bulk phase 1 International symbol: {}".format(data["international"]))
-print(f"Bulk phase 1 Lattice type: {sym.get_lattice_type()}")
-
-struct = Structure.from_file(filename2)
-sym = SpacegroupAnalyzer(struct)
-data = sym.get_symmetry_dataset()
-
-print("Bulk phase 2 Space group number: {}".format(data["number"]))
-print("Bulk phase 2 International symbol: {}".format(data["international"]))
-print(f"Bulk phase 2 Lattice type: {sym.get_lattice_type()}")
+def get_lattice_direction(prompt, log):
+    direction_input = input(prompt)
+    log.append(f"{prompt.strip()}: {direction_input}")
+    return [float(value.strip()) for value in direction_input.split(',')]
 
 
-# Manually develop interfaces/DW if you have ORs
-
-# Find the path of the current directory
-current_path = os.getcwd()
-    
-# initialize colorama
-init()
-
-#Creating planes
-
-# define lattice directions for slab1
-a_input = input(Fore.GREEN + Style.BRIGHT + "Enter three comma-separated values for bulk 1 lattice direction a: " + Style.RESET_ALL)
-a = [float(value.strip()) for value in a_input.split(',')]
-
-b_input = input(Fore.GREEN + Style.BRIGHT + "Enter three comma-separated values for bulk 1 lattice direction b: " + Style.RESET_ALL)
-b = [float(value.strip()) for value in b_input.split(',')]
-
-c_input = input(Fore.GREEN + Style.BRIGHT + "Enter three comma-separated values for bulk 1 lattice direction c: " + Style.RESET_ALL)
-c = [float(value.strip()) for value in c_input.split(',')]
-
- # define lattice directions for slab2
[…]
