### 2024-07-19 6c0465c
Update vasp2cif.py
