### 2024-07-19 cef9487
Delete vasp2cif.py
