+++ main.py
@@ -20,6 +20,8 @@ def main():
         '4': 'slab.py',
         '5': 'polarization.py'
         '6': 'supercell.py'
+        '7': 'vasp2cif.py'
+
 
     }
 
