+++ main.py
@@ -17,7 +17,7 @@ def display_dwbuilder_design():
     *********************************************
     *                                           *
     *               DWBuilder                   *
-    *                                           *
+    *                v2.0.0                     *
     *********************************************
     """
     print(design)
