+++ README.md
@@ -30,6 +30,60 @@ After installation, you can start the software by running the following command 
 ```
 > dwbuilder
 ```
+# Method for the development of FDW
+## R3m
+Ferroelectric materials with a rhombohedral space group of R3m exhibit three types of ferroelectric domain walls (FDWs): R71, R109, and R180. Due to neutrality and mechanical compatibility, these FDWs develop along specific planes. For the R180 FDW, the {1-10} plane is used, while for R71 and R109, the FDWs are determined by the sum of the two polarization vectors.
+
+- **R71 FDW**: Lies parallel to the diagonal (1-10) plane in the primitive cell.
+- **R180 FDW**: Developed along the {1-10} plane.
+- **R109 FDW**: The R109 FDW lies along the (100) plane, and FDWs are simply developed by stacking the primitive cell along this direction.
+
+To analyze the R71 and R180 FDWs, the pseudocubic rhombohedral unit cell is transformed into a 10-atom cell by rotating it 45° around the z-axis relative to the parent cubic unit cell. The lattice vectors used for this transformation are:
+
+```
+a=  [1, 1, 0], b=  [0, 0, 1], c= [1, -1, 0]
+
+```
+
+### R71 Domain Walls Development
+
+The R71 domain walls are developed by stacking along the normal direction to the [1-10] plane, which aligns with the a-axis in the rotated coordinate system. Additionally, the mirror symmtry was developed by tranforming [1-10] to [-110] and from [001] to [00-1].
+
+### Orientation Relationships (ORs) for R71 Domains
+
+For the R71 FDWs, the orientation relationships for both domains are:
+
+ ```
+a1=  [1, 1, 0], b1=  [0, 0, 1], c1= [1, -1, 0]
+a2=  [1, 1, 0], b2=  [0, 0, -1], c2= [-1, 1, 0]
+
+```
+### Orientation Relationships (ORs) for R180 Domains
+
+Similar to the R71 domain wall, the R180 {1-10} wall is also parallel to the diagonal plane in the primitive cell. For the R180 FDW:
+- A supercell is made by stacking the rotated 10-atom unit cell.
+- The final FDWs are developed by stacking along the [1-10] plane.
+
+ ```
+a1=  [1, 1, 0], b1=  [0, 0, 1], c1= [1, -1, 0]
+a2=  [-1, -1, 0], b2=  [0, 0, -1], c2= [1, -1, 0]
+
+```
+The transformation ensures that the polarization vectors in the two domains are oriented 180 degrees apart, achieving the necessary mirror symmetry for the R180 domain wall. 
+
+### Orientation Relationships (ORs) for 109 Domains
+
+ The R109 FDW lies along the (100) plane. For this FDW:
+ 
+- The domain walls are developed by stacking the primitive cell along the (100) direction.
+
+This straightforward stacking ensures that the polarization vectors align as required for the R109 domain wall configuration.
+
+ ```
+a1=  [1, 0, 0], b1=  [0, 1, 1], c1= [0, 0, 1]
+a2=  [1, 0, 0], b2=  [0, -1, 0], c2= [0,0 -1]
+
+```
 
 # Usage
[…]
