+++ scripts/bondanalysis.py
@@ -1,0 +1,204 @@
+#!/usr/bin/env python
+"""
+Script for bonding analysis in a VASP format structure.
+
+This script reads one or two VASP format structure files, determines the bonding between
+different elements based on a cutoff distance, and plots the bonding network
+and a histogram of bond lengths.
+
+Dependencies:
+- ase
+- matplotlib
+- networkx
+- numpy
+
+Usage:
+- Ensure the required dependencies are installed and the script is executed in a Python environment.
+- Run the script and provide the input file name when prompted.
+"""
+
+import ase.io
+import matplotlib.pyplot as plt
+import networkx as nx
+import numpy as np
+from ase.neighborlist import neighbor_list
+from matplotlib.lines import Line2D
+from collections import defaultdict
+
+def read_vasp_structure(file_name):
+    """Read the VASP structure file."""
+    return ase.io.read(file_name)
+
+def determine_bonding(structure, cutoff):
+    """Determine bonds based on a cutoff distance, considering periodic boundary conditions."""
+    indices_i, indices_j, offsets = neighbor_list('ijS', structure, cutoff)
+    bond_dict = defaultdict(list)
+    bond_lengths_dict = defaultdict(list)
+    for i, j, S in zip(indices_i, indices_j, offsets):
+        distance = structure.get_distance(i, j, mic=True)
+        if distance <= cutoff:
+            elem1 = structure[i].symbol
+            elem2 = structure[j].symbol
+            bond_type = tuple(sorted([elem1, elem2]))
+            bond_dict[bond_type].append((i, j, elem1, elem2))
+            bond_lengths_dict[bond_type].append(distance)
+    return bond_dict, bond_lengths_dict
+
+def plot_bonding_network(structure, bond_dict, shortest_bond=None, longest_bond=None, title_suffix=""):
+    """Plot the bonding network."""
+    G = nx.Graph()
+    positions = {}
+
+    for atom in structure:
+        G.add_node(atom.index, element=atom.symbol)
+        positions[atom.index] = atom.position[:2]
+
+    for bond_type, bonds in bond_dict.items():
+        color = np.random.rand(3,)  # Random color for each bond type
+        for bond in bonds:
+            i, j, elem1, elem2 = bond
+            if shortest_bond and ((i, j) == shortest_bond or (j, i) == shortest_bond):
+                G.add_edge(i, j, color='green', width=3.0)
+            elif longest_bond and ((i, j) == longest_bond or (j, i) == longest_bond):
+                G.add_edge(i, j, color='red', width=3.0)
+            else:
+                G.add_edge(i, j, color=color)
+
+    fig, ax = plt.subplots(figsize=(10, 10))
+    elements = list(set(atom.symbol for atom in structure))
+
+    colors = plt.cm.jet(np.linspace(0, 1, len(elements)))
+    color_map = {element: color for element, color in zip(elements, colors)}
+
+    node_colors = [color_map[G.nodes[node]['element']] for node in G.nodes]
+    edge_colors = [G[u][v]['color'] for u, v in G.edges]
+
+    nx.draw(G, pos=positions, node_color=node_colors, with_labels=True, ax=ax, node_size=500, font_size=10, edge_color=edge_colors)
+
+    legend_elements = [Line2D([0], [0], marker='o', color='w', markerfacecolor=color, markersize=10, label=element)
+                       for element, color in color_map.items()]
+    
+    bond_legend_elements = [Line2D([0], [0], color=color, lw=2, label='-'.join(bond_type)) for bond_type, color in zip(bond_dict.keys(), edge_colors)]
[…]
