+++ README.md
@@ -56,24 +56,22 @@ a=[1, 1, 0], b=[0, 0, 1], c=[1, -1, 0]
 The R71 domain walls are developed by stacking along the normal direction to the [1-10] plane, which aligns with the a-axis in the rotated coordinate system. Additionally, the mirror symmtry was developed by tranforming [1-10] to [-110] and from [001] to [00-1].
 
 For the R71 FDWs, the orientation relationships for both domains (D1 and D2) are:
-```
 
          [ 1 0  1 ]         [ 1  0 -1 ]          
     D1 = [ 1 0 -1 ]    D2 = [ 1  0  1 ]    
          [ 0 1  0 ]         [ 0 -1  0 ]  
 
- ```
 **R180**
 
 Similar to the R71 domain wall, the R180 {1-10} wall is also parallel to the diagonal plane in the primitive cell. For the R180 FDW:
 - A supercell is made by stacking the rotated 10-atom unit cell.
 - The final FDWs are developed by stacking along the [1-10] plane.
+- 
+         [ 1 0  1 ]         [ -1  0   1 ]          
+    D1 = [ 1 0 -1 ]    D2 = [ -1  0  -1 ]    
+         [ 0 1  0 ]         [  0 -1   0 ]  
 
- ```
-a1=[1, 1, 0], b1=[0, 0, 1], c1=[1, -1, 0]
-a2=[-1, -1, 0], b2=[0, 0, -1], c2=[1, -1, 0]
 
-```
 The transformation ensures that the polarization vectors in the two domains are oriented 180 degrees apart, achieving the necessary mirror symmetry for the R180 domain wall. 
 
 **R109**
