+++ README.md
@@ -90,6 +90,9 @@ This straightforward stacking ensures that the polarization vectors align as req
          [ 0  1  1 ]         [  0  0 -1 ]  
 
 ```
+
+## Reference: Taherinejad M, Vanderbilt D, Marton P, Stepkova V, Hlinka J. Bloch-type domain walls in rhombohedral BaTiO3. Physical Review B. 2012 Oct 22;86(15):155138.
+
 ## R3c
 Similar to the R3m, the R3c structure also possesses three FDWs: R71, R109, and R180.  The R3c domain walls are developed by converting the reference R3c structure into a pseudocubic cell. This is done by cutting along the ```(1-11)``` plane, which is also the polarization axis.
 
@@ -121,6 +124,8 @@ The R71 , R109 , and R180 domain walls are found to adopt ```(001)```, ```(110)`
 ```
 The R109 domain walls are formed by stacking the D1 and D2 domains along the ```[001]```, while R71 and R180 are stacked along the ```[110]``` and ```[-110]``` directions, respectively. 
 
+## Reference: Wang Y, Nelson C, Melville A, Winchester B, Shang S, Liu ZK, Schlom DG, Pan X, Chen LQ. BiFeO 3 domain wall energies and structures: a combined experimental and density functional theory+ U study. Physical review letters. 2013 Jun 24;110(26):267601.
+
 ## P4mm
 In the P4mm space group, with the polarization direction along the `[001]` direction, there are two types of domain walls allowed and developed in `dwbuilder`:
 
@@ -147,6 +152,8 @@ The transformation matrix used for both domain walls are given as:
 ```
 
 The T180 and T90 FDWs are develed by stacking along the ```[010]``` and ```[101]``` direction.  
+
+## Reference: Meyer B, Vanderbilt D. Ab initio study of ferroelectric domain walls in PbTiO 3. Physical Review B. 2002 Mar 1;65(10):104111.
 
 ## P6_3cm
 
@@ -183,7 +190,7 @@ By providing the necessary input structures, these scripts will generate the res
 
 **Note**: You do not explicitly need to run these scripts directly. `dwbuilder` automatically detects the space group and, based on the space group, prompts users to provide details of the input structure and parameters.
 
-
+## Reference: Kumagai Y, Spaldin NA. Structural domain walls in polar hexagonal manganites. Nature communications. 2013 Feb 26;4(1):1540.
 
 # Usage
 To execute the code, one should run the dwbuilder command from the command line. Upon running, the user will be prompted to input the name of the input structure file, as well as the domain wall angle and size (expressed in number of unit cells). This script builds the domain wall structure based on the space group of the input structure. For structures with the R3m or P4mm space groups, the polarization directions are assumed to align with the ```[001]``` direction. For structures with the R3c space group, a pseudo-cubic structure is first constructed by translating the input rhombohedral axes along ```a=[101], b=[-111], and c=[0-11]```. In the translated axis , the polarization direction is assumed to align with the ```[-110]``` direction.  To construct 109 and 71 domain walls, each crystal system is translated according to specific orientation relationships in order to meet the polarization angle requirement.
