+++ README.md
@@ -244,6 +244,15 @@ The Pmc2_1 and Amm2 space groups belong to the orthorhombic crystal system. Mate
    - Based on the user selection, `DWBuilder` applies the rotation matrix to develop the domain walls.
    - The rotation matrix is applied around the z-axis.
 
+ ```
+
+              [ cos(γ)  -sin(γ)  1 ]                    
+      Rz(γ) = [ sin(γ)   cos(γ) -1 ]        
+              [ 0         0      0 ]               
+
+```
+
+
 3.  **Polar Axis Adjustment**:
    - `DWBuilder` ensures that the polar axis does not lie along the z-axis before applying the rotation matrix.
    - If the user selects the a-axis as the polar axis, the structure is kept as is.
@@ -255,16 +264,7 @@ The Pmc2_1 and Amm2 space groups belong to the orthorhombic crystal system. Mate
               [ 0  1  0 ]               [  1  0   0 ]  
 ```
     
-    - After aligning the a-axis as the polar axis, `DWBuilder` applies the rotation matrix around the z-axis using the following relation:
-
-
-```
-
-              [ cos(γ)  -sin(γ)  1 ]                    
-      Rz(γ) = [ sin(γ)   cos(γ) -1 ]        
-              [ 0         0      0 ]               
-
-```
+    - After aligning the a-axis as the polar axis, `DWBuilder` applies the rotation matrix around the z-axis by applying `(Rz(γ))`.
 
 # Usage
 To execute the code, one should run the dwbuilder command from the command line. Upon running, the user will be prompted to input the name of the input structure file, as well as the domain wall angle and size (expressed in number of unit cells). This script builds the domain wall structure based on the space group of the input structure. For structures with the R3m or P4mm space groups, the polarization directions are assumed to align with the ```[001]``` direction. For structures with the R3c space group, a pseudo-cubic structure is first constructed by translating the input rhombohedral axes along ```a=[101], b=[-111], and c=[0-11]```. In the translated axis , the polarization direction is assumed to align with the ```[-110]``` direction.  To construct 109 and 71 domain walls, each crystal system is translated according to specific orientation relationships in order to meet the polarization angle requirement.
