+++ README.md
@@ -264,7 +264,11 @@ The Pmc2_1 and Amm2 space groups belong to the orthorhombic crystal system. Mate
          Rb = [ 1  0  0 ]         Rz =  [  0  0   1 ]    
               [ 0  1  0 ]               [  1  0   0 ]  
 ```
-    
+
+## - _**A note on the selection of domain wall termination sites:**_
+
+`DWBuilder` develops domain wall supercells by generating and stacking two different domains, with the termination of the domains depending on the input unit cell termination. For example, PbTiO\(_3\) in the P4mm space group can have TiO- or PbO-terminated domain wall structures. To develop both variants, the user needs to define the input unit cell structure such that in one variant, the Pb atom is in the center, leading to PbO-terminated domain walls, while in the other, the Ti atom is in the center, resulting in TiO-terminated domain walls. Additionally, for complex systems, `DWBuilder` includes a function to add one or two layers by defining the domain size in fractional form (e.g., 2.5), providing precise control over the domain wall structure and thickness.
+
 # Usage
 To execute the code, one should run the dwbuilder command from the command line. Upon running, the user will be prompted to input the name of the input structure file, as well as the domain wall angle and size (expressed in number of unit cells). This script builds the domain wall structure based on the space group of the input structure. For structures with the R3m or P4mm space groups, the polarization directions are assumed to align with the ```[001]``` direction. For structures with the R3c space group, a pseudo-cubic structure is first constructed by translating the input rhombohedral axes along ```a=[101], b=[-111], and c=[0-11]```. In the translated axis , the polarization direction is assumed to align with the ```[-110]``` direction.  To construct 109 and 71 domain walls, each crystal system is translated according to specific orientation relationships in order to meet the polarization angle requirement.
 
