### 2024-07-22 df5ca6a
Delete examples/P4mm/LOGFILE.txt
