+++ README.md
@@ -260,9 +260,9 @@ The Pmc2_1 and Amm2 space groups belong to the orthorhombic crystal system. Mate
    - After aligning the a-axis as the polar axis, `DWBuilder` applies the rotation matrix around the c-axis by applying `(Rc(γ))`.
 
 ```
-       [ 1  0  0 ]       [ 0  0  1 ]               [  0  1   0 ]          
- Ra =  [ 0  1  0 ]  Rb = [ 1  0  0 ]         Rc =  [  0  0   1 ]    
-       [ 0  0  1 ]       [ 0  1  0 ]               [  1  0   0 ]  
+       [ 1  0  0 ]       [ 0  0  1 ]        [ 0  1  0 ]          
+ Ra =  [ 0  1  0 ]  Rb = [ 1  0  0 ]  Rc =  [ 0  0  1 ]    
+       [ 0  0  1 ]       [ 0  1  0 ]        [ 1  0  0 ]  
 ```
 
 3.  **O120 Head-to-Head/Tail-to-Tail and Head-to-Tail domain walls**:
