### 2024-07-24 #6895 @mzkhalid039
> I have read the paper, the documentation and the codes. The code builds domain wall structures, which is indeed potentially useful for researchers working in the field of (anti)ferroelectric materials. The paper is clear and well written. But I find there are a few issues with the documentation and the coding, which makes the code less practical. I think many improvements are needed before the publication.

Thank you for your time and valuable suggestions. We have incorporated your feedback into the documentation and code, and have released the updated version as ```DWBuilder v2.0.0```.

> * There is no documenation of the method being used internally in the code, nor what kind of domain structure is built. There is no unique way to build domain structures, and using different ways, the strcuture will also vary. For researchers, it is essential to understand what is done.

We have updated the README file with a section titled ```Method for the development of FDW```. This section explains the methodology behind the development of ferroelectric/ferroelastic domain walls.

> * The installation script has some bugs:
>   ```
>   pip install .
>   ```
>
>
>
>
>
>
>
>
>
>
>
>   runs, but the scripts installed won't work as expected
>   This is because in setup.py,
>   ```
>       entry_points={
>           'console_scripts': [
>               'dwbuilder = scripts.dwbuilder:main',
>               'dbuilder = scripts.dbuilder:main',
>               'hibuilder = scripts.hibuilder:main',
>               'polarization = scripts.polarization:main',
>           ]
>
>   ```
>
>
>
>
>
>
>
>
>
>
>
>   But there is no main function in the script/dwbuilder.py file.
>   Same for the other scripts here. When running the scripts, there will be this error:
>   ```
>   AttributeError: module 'scripts.dwbuilder' has no attribute 'main'
>   ```

We have fixed this issue by including the ```main.py```  file in the root directory and updating the ```setup.py``` file accordingly:

```
    entry_points={
        'console_scripts': [
            'dwbuilder = main:main',
        ]
```

> * Related to the issue above, the variables and functions in scripts/*.py are in the global namespace, not inside  functions/classes. Actually I didn't see any function or class defined in the code.   This is not the best practice, which makes the code less readable and maintainable.  And importing the module will execute all the codes.

Thank you for the excellent suggestion. We have completely rewritten the code for ```dwbuilder.py``` and ```dbuilder.py```, organizing them into functions and classes. Additionally, we have created a dictionary for the transformation matrices used in the development of domain walls. I agree, this makes it easy to read and maintain the code.

> * When the package is installed, the commands to run are dwbuilder,dbuilder, hibuilder, and polarization instead of dwbuilder.py, ....
>   This is not clear enough in the README.md file.

We have fixed this issue and clarified the instructions in the README.md file. Now, users only need to run the ```dwbuilder``` command from the terminal, which will display a home screen and prompt the user to select a specific script to run.

> * The functionality is not complete enough to be used in practice. When building the domain walls, the Warning message shows:
>   ```
>   Warning: This code might create domain wall artifacts, like missing or duplicate atoms, at the domain wall. Therefore, manual adjustment may be needed.
>   ```
>
>
>
>
>
>
>
>
>
>
>
>   I can see that the problem is because in the implementation, the domain structure is simply built by joining two rotated slabs. For example, in dwbuilder.py in the following lines:
>   ```
>       if angle_selection == 1 or angle_selection == 4:
>           slab1 = cut(BFO, a=[0.0, 1, 0], b=[1, 0, 0], c=[0, 0, domain_size])
>           rotate(slab1, slab1.cell[0], (0,1,0), slab1.cell[1], (1,0,0))
>   ```
>
>
>
>
>
>
>
>
>
>
>
>   With this strategy, it seems that the code will indeed generate duplicate or missing atoms quite easily, as there is no guarantee at all that no conflicts will be generated as interface with the rotation.
>   This is a serious problem for building domain wall structures, which hinders the code from being useful in practise.  There are ways to avoid these problems by adding (a) detecting the duplicated or missing atoms (b)shifting the atoms to its periodic images.

We have addressed the issue by adding a functionality that prompts the user to input a cutoff limit to remove close atoms during the development of domain walls. This feature removes all atoms at or below the cutoff limit.

Additionally, we have fixed the issue with the development of domain walls in hexagonal manganites, which previously resulted in domain wall structures with missing atoms. We have now included new scripts, ```ndw.py``` and ```cdw.py```, which generate perfect domain wall structures without duplicating or missing atoms.

> * It is not possible to specify the domain wall size and the supercell size as m×n×k or even in matrix form, so only n_n_n shaped domain wall/supercell structures are possible.  If the structure is built for first-principles calculation, it usually requires  to minimize the  number of atoms.

We have included a new script, ```rotationmatrix.py```, which allows the definition of the entire rotation matrix in the form of m×n×k. However, we have retained the ```dwbuilder.py``` supercell in the form of n_n_n. This choice was made to keep the development of domain walls simple and straightforward.

> * It is not clear (a) whether there are multiple possible domain walls (b) if so, how to choose from the possibilities. For example, the 180 degree domain wall of PbTiO3 can be Ti-O centered or Pb-O centered, and the wall can be perpendicular or parallel to the supercell direction.

>   While in the code, the assumed polarization direction is documenated. Thus for the user, it is possible to shift and rotate the input structure to adapt to the code. But this is not simple nor clearly documented, as it requires the knowledge of the internals of how the domain wall is assembled.

The ```dwbuilder.py``` script constructs a domain wall supercell by using the input unit cell structure as a base. To create different termination sites, users need to adjust the input unit cell structure so that various termination sites are exposed during the supercell construction. We have provided an example of this process using PbTiO3 in the P4mm example folder.

Additionally, we have included a note about this in the README.md file. A methodology section has been added to the README.md file to help users understand the mechanism behind the development of domain walls.

> * When rebuilding domain wall structures in the same directory, the FileExistsError is raised. It can be made more user-friendly.

We have fixed this issue. Now, if a folder with the same name already exists, the system will automatically generate a copy folder.

> * The "DWBuilder.egg-info" directory should not be included in the git repo.

It has been removed.

> * There is no atomated test. Actually, the way the code is designed makes it very difficult for unit tests, as there is no "unit".

I am unsure about how to run the automated test. However, we have tested the domain walls with different space groups using the referenced material systems provided in the examples folder of the script. The code successfully generates the expected domain wall structures.
