+++ README.md
@@ -1,5 +1,4 @@
 ![GitHub release (latest by date)](https://img.shields.io/github/v/release/mzkhalid039/DWBuilder)
-![Python Version](https://img.shields.io/badge/python-3.11-blue)
 [![status](https://joss.theoj.org/papers/eeae372c75342085351f92af4ca165db/status.svg)](https://joss.theoj.org/papers/eeae372c75342085351f92af4ca165db)
 
 
