### 2024-08-18 #6895 @Anshuman5
This work presents a highly valuable tool for the materials science community by providing a package that automates the creation of domain walls and interfaces. I recommend the manuscript for acceptance after addressing the following comments:

**Comment 1**: Please modify the `README` file. Specifically, remove the section titled "Method for the Development of FDW" and create a separate markdown file for this content. Provide a link to this file at the end of the README. The suggested order for the README is as follows:
- Overview
- Installation Guide
- Usage
- Examples
- Method Development (only link)
- References

**Comment 2**: The paper should include usage examples for both domain walls and heterogeneous interfaces. Additionally, please incorporate figures corresponding to these examples in the paper. This section is currently missing from the paper and is only included in the `README` on GitHub.

**Comment 3**: There is inconsistency in the documentation across the codebase. While some scripts (e.g., `bondanalysis.py`) are documented, others (e.g., `cdw`, `dbuilder`, `polarization`, etc) lack comments for their classes and functions. Please ensure that all classes and functions in the Python files are properly documented with comments.
