+++ paper/paper.md
@@ -40,6 +40,37 @@ Currently, the script cannot predict the ORs that would result in a low lattice 
 
 The ``DWBuilder`` code is designed to automate the creation of atomic interfaces and domain wall structures, allowing researchers to focus on optimizing and studying material behavior and properties. The structures generated by this code are compatible with both first-principles and second-principles calculations. The code provides ample functionality to support practical research tasks while remaining lightweight and well-documented. This allows junior researchers with minimal or no assistance to easily install, use, and understand the code.
 
+# Example:
+
+The below example demonstrates the functionality and steps of the `dwbuilder.py` script. Running the `dwbuilder` script will perform the following steps:
+
+1. **User Input**:
+    - Specify the script to run on the input structure.
+    - Input the VASP format structure.
+
+2. **Space Group Determination**:
+    - The script determines the space group of the provided structure.
+
+3. **Allowed Domain Wall Types**:
+    - Based on the determined space group, the script identifies the allowed domain wall types.
+
+4. **Domain Wall and Supercell Size**:
+    - User inputs the desired domain wall size.
+    - User specifies the supercell size for the domain wall structures.
+
+5. **Final Domain Wall Structures**:
+    - The script generates the final domain wall structures.
+    - It prints any lattice misfit information.
+
+ ![Structure of the `DWBuilder` package.\label{fig:2}](dwbuilder_example){width="100%"}
+
+Figure 2 illustrates an example of P4mm PbTiO$_3$ and hexagonal manganite YMnO$_3$ domain wall structures.
+
+For proper ferroelectrics, such as perovskites (PbTiO$_3$, BaTiO$_3$, KNbO$_3$, etc.), the user only needs to define the primitive unit cell structure. The `DWBuilder` automatically determines the space group and constructs the respective domains and domain wall structures.
+
+In certain cases, if the space group of the input structure does not match the defined space group type of the domain walls, the user can manually build the domains. This can be done by defining the transformation matrix of each domain using either `hibuilder.py` or by selecting the manual option for domain selection in `dwbuilder.py`.
+
+For hexagonal manganites, both domain structures are required to build domain wall structures, as illustrated in Figure 2.
 
 # Acknowledgements
 
