+++ paper/paper.md
@@ -40,7 +40,7 @@ Currently, the script cannot predict the ORs that would result in a low lattice 
 
 The ``DWBuilder`` code is designed to automate the creation of atomic interfaces and domain wall structures, allowing researchers to focus on optimizing and studying material behavior and properties. The structures generated by this code are compatible with both first-principles and second-principles calculations. The code provides ample functionality to support practical research tasks while remaining lightweight and well-documented. This allows junior researchers with minimal or no assistance to easily install, use, and understand the code.
 
-# Example:
+# Example 1:
 
 The below example demonstrates the functionality and steps of the `dwbuilder.py` script. Running the `dwbuilder` script will perform the following steps:
 
@@ -62,8 +62,7 @@ The below example demonstrates the functionality and steps of the `dwbuilder.py`
     - The script generates the final domain wall structures.
     - It prints any lattice misfit information.
 
- ![ ![Illustration of domain wall structures in PbTiO$_3$ and YMnO$_3$ using the `dwbuilder.py` script. PbTiO$_3$, a proper ferroelectric belonging to the tetragonal P4mm space group, exhibits a polar axis along the c-axis in the primitive unit cell structure, as shown in (a). The `DWBuilder` script automatically determines the space group and constructs the (b) T90 and (c) T180 domain wall structures. For YMnO$_3$, with two domain structures, (d) am and (e) bp, both exhibiting polarization along the $\pm c$-axis, the script generates (f) neutral domain walls by stacking the domains along the b-axis and (g) charged domain walls by stacking them along the c-axis.\label{fig:2}](dwbuilder_example){width="100%"}
-\label{fig:2}](dwbuilder-example){width="100%"}
+ ![ ![Illustration of domain wall structures in PbTiO$_3$ and YMnO$_3$ using the `dwbuilder.py` script. PbTiO$_3$, a proper ferroelectric belonging to the tetragonal P4mm space group, exhibits a polar axis along the c-axis in the primitive unit cell structure, as shown in (a). The `DWBuilder` script automatically determines the space group and constructs the (b) T90 and (c) T180 domain wall structures. For YMnO$_3$, with two domain structures, (d) am and (e) bp, both exhibiting polarization along the $\pm c$-axis, the script generates (f) neutral domain walls by stacking the domains along the b-axis and (g) charged domain walls by stacking them along the c-axis.\label{fig:2}](dwbuilder_example){width="100%"}\label{fig:2}](dwbuilder-example){width="100%"}
 
 Figure 2 illustrates an example of P4mm PbTiO$_3$ and hexagonal manganite YMnO$_3$ domain wall structures.
 
@@ -72,6 +71,19 @@ For proper ferroelectrics, such as perovskites (PbTiO$_3$, BaTiO$_3$, KNbO$_3$, 
 In certain cases, if the space group of the input structure does not match the defined space group type of the domain walls, the user can manually build the domains. This can be done by defining the transformation matrix of each domain using either `hibuilder.py` or by selecting the manual option for domain selection in `dwbuilder.py`.
 
 For hexagonal manganites, both domain structures are required to build domain wall structures, as illustrated in Figure 2.
+
+# Example 2:
+
+The following example illustrates the use of the `hibuilder.py` script. This script performs the following steps:
+
+1. Inputs the primitive unit cell structures of the domain or bulk structures.
+2. Defines the compatible transformation matrices for both input domains or bulk structures.
+3. `hibuilder.py` develops the transformed bulk structures.
+4. Stacks the transformed bulk structures to build the final interface and calculates the lattice and angular misfit.
+
[…]
