+++ paper/paper.md
@@ -79,7 +79,7 @@ The following example illustrates the use of the `hibuilder.py` script. This scr
 3. `hibuilder.py` develops the transformed bulk structures.
 4. Stacks the transformed bulk structures to build the final interface and calculates the lattice and angular misfit.
 
-Figure 3 illustrates an example of the interface structure between Fe and Fe$_2$Al$_5$. This example is taken from the study by [@khalid2020ab]. 
+Figure 3 illustrates an example of the interface structure between Fe and Fe$_2$Al$_5$. This example is taken from the study by [@khalid2021first]. 
 
 ![ ![Illustration of domain wall structures in PbTiO$_3$ and YMnO$_3$ using the `dwbuilder.py` script. PbTiO$_3$, a proper ferroelectric belonging to the tetragonal P4mm space group, exhibits a polar axis along the c-axis in the primitive unit cell structure, as shown in (a). The `DWBuilder` script automatically determines the space group and constructs the (b) T90 and (c) T180 domain wall structures. For YMnO$_3$, with two domain structures, (d) am and (e) bp, both exhibiting polarization along the $\pm c$-axis, the script generates (f) neutral domain walls by stacking the domains along the b-axis and (g) charged domain walls by stacking them along the c-axis.\label{fig:2}](dwbuilder_example){width="100%"}\label{fig:2}](dwbuilder-example){width="100%"}
 
