+++ paper/paper.md
@@ -84,7 +84,7 @@ Figure 3 illustrates an example of the interface structure between Fe and Fe$_2$
 ![ ![Illustration of domain wall structures in PbTiO$_3$ and YMnO$_3$ using the `dwbuilder.py` script. PbTiO$_3$, a proper ferroelectric belonging to the tetragonal P4mm space group, exhibits a polar axis along the c-axis in the primitive unit cell structure, as shown in (a). The `DWBuilder` script automatically determines the space group and constructs the (b) T90 and (c) T180 domain wall structures. For YMnO$_3$, with two domain structures, (d) am and (e) bp, both exhibiting polarization along the $\pm c$-axis, the script generates (f) neutral domain walls by stacking the domains along the b-axis and (g) charged domain walls by stacking them along the c-axis.\label{fig:2}](dwbuilder_example){width="100%"}\label{fig:2}](dwbuilder-example){width="100%"}
 
 
- ![ ![Illustration of the `hibuilder.py` script used to build the interface structures between (a) Fe and (b) Fe$_2$Al$_5$. The transformed bulk structures are shown in (c) and (d) respectively. The final interface structure is created by stacking the transformed bulk phases so that Fe$[22-2] is parallel to Fe$_2$Al$_5$[100]. (e) The interface structure can be further optimized by defining the optimal distance between the two bulk phases.  \label{fig:3}](hibuild-example){width="100%"}
+ ![ ![Illustration of the `hibuilder.py` script used to build the interface structures between (a) Fe and (b) Fe$_2$Al$_5$. The transformed bulk structures are shown in (c) and (d) respectively. The final interface structure is created by stacking the transformed bulk phases so that Fe$[22-2] is parallel to Fe$_2$Al$_5$[100]. (e) The interface structure can be further optimized by defining the optimal distance between the two bulk phases.\label{fig:3}](hibuild-example){width="100%"}\label{fig:3}]
 
 # Acknowledgements
 
