### 2024-08-23 cce4d57
Update polarization.py
