+++ scripts/supercell.py
@@ -1,3 +1,35 @@
+#!/usr/bin/env python
+
+"""
+Script: supercell.py
+Description:
+    This script generates a supercell from a given VASP POSCAR file. The user specifies the supercell dimensions 
+    along the a, b, and c crystallographic directions. The resulting supercell structure is then saved to a new 
+    VASP POSCAR file.
+
+Key Features:
+    - Prompts the user for input: file name of the VASP POSCAR file and supercell dimensions.
+    - Validates the existence of the input file in the current directory.
+    - Constructs the supercell using the specified dimensions.
+    - Saves the generated supercell to a new POSCAR file in the current directory.
+
+Usage:
+    - Run the script by executing `python supercell.py` and follow the on-screen prompts.
+
+Dependencies:
+    - ASE (Atomic Simulation Environment): For handling atomic structures, reading/writing POSCAR files, and creating supercells.
+    - Numpy: For matrix operations during supercell construction.
+    - OS: For handling file paths and checking file existence.
+
+Functions:
+    - get_user_input(): Prompts the user for the input file name and supercell dimensions, and validates the file's existence.
+    - build_supercell(structure, a, b, c): Constructs the supercell based on the user-provided dimensions.
+    - main(): The main function that coordinates user input, supercell construction, and file output.
+
+Output:
+    - The script outputs the supercell structure to a new POSCAR file in the current directory, with a modified file name indicating the supercell.
+"""
+
 import numpy as np
 from ase import io
 from ase.build import make_supercell
