+++ README.md
@@ -29,6 +29,149 @@ After installation, you can start the software by running the following command 
 ```
 > dwbuilder
 ```
+
+# Usage
+To execute the code, one should run the dwbuilder command from the command line. Upon running, the user will be prompted to input the name of the input structure file, as well as the domain wall angle and size (expressed in number of unit cells). This script builds the domain wall structure based on the space group of the input structure with the assumed polarization directions as indicated below. 
+
+```
+R3c (Polarization direction (1-11))
+R3m (Polarization direction (001))
+P4mm (Polarization direction (001))
+p6_3cm (Polarization direction (-1-10))
+Pnma (Polarization direction (110))
+Pmc2_1 (user-defined)
+Amm2 (user-defined)
+```
+
+However, if you are unsure about the polarization dirction, you can run [polarization.py](scripts/polarization.py) to determine the polarization direction. This script uses the point charge model and you need to have both optimized ```CONTCAR``` and high symmtry structure ```POSCAR``` files to determine the polarization.
+
+## Example of using [dwbuilder.py](scripts/dwbuilder.py)
+_ex:_
+
+```
+> dwbuilder
+
+    ***************************************************
+    *                                                 *
+    *                Welcome to DWBuilder             *
+    *                                                 *
+    ***************************************************
+    *        Hey, you must know what you are doing.   *
+    *  Otherwise, you might get wrong results.        *
+    ***************************************************
+    *     Core Developer: M.Z.Khalid                  *
+    *     Main Contributors: S.M.Selbach              *
+    *     Email: zeeshan.khalid039@gmail.com          *
+    ***************************************************
+    
+Select a script to run:
+1: dwbuilder.py                6: supercell.py
+2: dbuilder.py                 7: vasp2cif.py
+3: hibuilder.py                8: bondanalysis.py
+4: slab.py                     9: rotationmatrix.py
+5: polarization.py             
+Enter the number of the script to run: 1
+Enter the input file name (with extension): BiFeO3_R3c.vasp
+Space group number: 161
+International symbol: R3c
+Lattice type: rhombohedral
+Select the domain wall angle:
+R180 - R180
+R71 - R71
+R109 - R109
+ALL - ALL
+ALL
+Enter the domain wall size (in number of unit cells): 3
+Enter the size of the supercell along the a direction: 1
+Enter the size of the supercell along the b direction: 1
+Enter the size of the supercell along the c direction: 1
+Enter the cutoff distance in angstroms to remove close atoms: 0
+Lattice strain for R109 DW:
+strain along a (%): 0.00
+strain along b (%): 0.00
+strain along c (%): 0.00
+Lattice strain for R71 DW:
+strain along a (%): 0.00
+strain along b (%): 0.00
+strain along c (%): 0.00
+Lattice strain for R180 DW:
+strain along a (%): 0.00
+strain along b (%): 0.00
+strain along c (%): 0.00
+Domain wall structures and supercells created successfully! Log file written to LOGFILE.txt
+```
+
+Once you have developed the desired domain wall structure, you can visualize in vesta or ase to further refine domain wall artifacts. 
+
+## Example of using [hibuilder.py](scripts/hibuilder.py)
+In this example, I have reproduced an orientation relationship reported in the following article [CMS](https://www.sciencedirect.com/science/article/pii/S0927025621000446). 
+
+```
+> dwbuilder
+
+    ***************************************************
+    *                                                 *
+    *                Welcome to DWBuilder             *
+    *                                                 *
+    ***************************************************
+    *        Hey, you must know what you are doing.   *
[…]
