### 2024-08-23 #6895 @mzkhalid039
> This work presents a highly valuable tool for the materials science community by providing a package that automates the creation of domain walls and interfaces. I recommend the manuscript for acceptance after addressing the following comments:

Thank you very much for taking the time to review our paper and software. We have made updates to both the paper and the README.md file based on your suggestions and comments. Please review our responses below, along with the updated paper and script files, to see the changes we have made.

> **Comment 1**: Please modify the `README` file. Specifically, remove the section titled "Method for the Development of FDW" and create a separate markdown file for this content. Provide a link to this file at the end of the README. The suggested order for the README is as follows:
>
> * Overview
> * Installation Guide
> * Usage
> * Examples
> * Method Development (only link)
> * References

Unfortunately, we do not have a separate website domain to publish our software documentation, we are limited to presenting the methodology on GitHub. However, to maintain the flow of practical installation and usage instructions, we have moved the methodology section to the end, as per your suggestion.

> **Comment 2**: The paper should include usage examples for both domain walls and heterogeneous interfaces. Additionally, please incorporate figures corresponding to these examples in the paper. This section is currently missing from the paper and is only included partly in the `README` on GitHub.

We have updated the paper with examples of domain walls and heterogenous interfaces.

> **Comment 3**: There is inconsistency in the documentation across the codebase. While some scripts (e.g., `bondanalysis.py`) are documented, others (e.g., `cdw`, `dbuilder`, `polarization`, etc) lack comments for their classes and functions. Please ensure that all classes and functions in the Python files are properly documented with comments.

We have updated all the scripts to include comments throughout, covering everything from classes and functions to practical implementation and dependencies.
