### 2024-08-23 #6895 @rkurchin
Congrats, @mzkhalid039, we're nearly finished! I'll send editorial comments on the paper soon; in the meantime, you can start working on the author tasks listed above and send me the relevant info (version tag and DOI) once you have them.
