### 2024-08-23 #6895 @rkurchin
@mzkhalid039 please check on the suggestions above and add DOI's for everything you can
