+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={3},
   pages={157--173},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi:{https://doi.org/10.1038/s41578-021-00375-z}
 }
 
 @article{catalan2012domain,
@@ -17,7 +18,8 @@
   number={1},
   pages={119},
   year={2012},
-  publisher={APS}
+  publisher={APS},
+  doi:{https://doi.org/10.1103/RevModPhys.84.119}
 }
 
 @article{meier2015functional,
@@ -28,7 +30,8 @@
   number={46},
   pages={463003},
   year={2015},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi:{DOI 10.1088/0953-8984/27/46/463003}
 }
 
 @article{bednyakov2018physics,
@@ -39,7 +42,8 @@
   number={1},
   pages={1--11},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi:{https://doi.org/10.1038/s41524-018-0121-8}
 }
 
 @article{whyte2015diode,
@@ -50,7 +54,8 @@
   number={1},
   pages={1--5},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi:{https://doi.org/10.1038/ncomms8361}
 }
 
 @article{mundy2017functional,
@@ -61,7 +66,8 @@
   number={6},
   pages={622--627},
   year={2017},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi:{https://doi.org/10.1038/nmat4878}
 }
 
 @article{sharma2017nonvolatile,
@@ -83,7 +89,8 @@
   number={26},
   pages={262903},
   year={2020},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi:{https://doi.org/10.1063/5.0009185}
 }
 
 @article{smaabraaten2020domain,
@@ -94,7 +101,8 @@
   number={3},
   pages={033159},
   year={2020},
-  publisher={APS}
+  publisher={APS},
+  doi:{https://doi.org/10.1103/PhysRevResearch.2.033159}
 }
 
 @article{smaabraaten2018charged,
@@ -105,7 +113,8 @@
   number={11},
   pages={114405},
   year={2018},
-  publisher={APS}
+  publisher={APS},
+  doi:{https://doi.org/10.1103/PhysRevMaterials.2.114405}
 }
 
 @article{ederer2005effect,
@@ -116,7 +125,8 @@
   number={25},
   pages={257601},
   year={2005},
-  publisher={APS}
+  publisher={APS},
+  doi:{https://doi.org/10.1103/PhysRevLett.95.257601}
 }
 
 @article{wang2013bifeo,
@@ -127,7 +137,8 @@
   number={26},
   pages={267601},
   year={2013},
-  publisher={APS}
+  publisher={APS},
+  doi:{https://doi.org/10.1103/PhysRevLett.110.267601}
 }
 
 @article{wang2003epitaxial,
@@ -138,7 +149,8 @@
   number={5613},
   pages={1719--1722},
   year={2003},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi:{10.1126/science.1080615}
 }
 
 @article{taherinejad2012bloch,
@@ -149,7 +161,8 @@
   number={15},
   pages={155138},
   year={2012},
-  publisher={APS}
[…]
