+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={157--173},
   year={2022},
   publisher={Nature Publishing Group UK London},
-  doi:{https://doi.org/10.1038/s41578-021-00375-z}
+  doi={https://doi.org/10.1038/s41578-021-00375-z}
 }
 
 @article{catalan2012domain,
@@ -19,7 +19,7 @@
   pages={119},
   year={2012},
   publisher={APS},
-  doi:{https://doi.org/10.1103/RevModPhys.84.119}
+  doi={https://doi.org/10.1103/RevModPhys.84.119}
 }
 
 @article{meier2015functional,
@@ -31,7 +31,7 @@
   pages={463003},
   year={2015},
   publisher={IOP Publishing},
-  doi:{DOI 10.1088/0953-8984/27/46/463003}
+  doi={DOI 10.1088/0953-8984/27/46/463003}
 }
 
 @article{bednyakov2018physics,
@@ -43,7 +43,7 @@
   pages={1--11},
   year={2018},
   publisher={Nature Publishing Group},
-  doi:{https://doi.org/10.1038/s41524-018-0121-8}
+  doi={https://doi.org/10.1038/s41524-018-0121-8}
 }
 
 @article{whyte2015diode,
@@ -55,7 +55,7 @@
   pages={1--5},
   year={2015},
   publisher={Nature Publishing Group},
-  doi:{https://doi.org/10.1038/ncomms8361}
+  doi={https://doi.org/10.1038/ncomms8361}
 }
 
 @article{mundy2017functional,
@@ -67,7 +67,7 @@
   pages={622--627},
   year={2017},
   publisher={Nature Publishing Group},
-  doi:{https://doi.org/10.1038/nmat4878}
+  doi={https://doi.org/10.1038/nmat4878}
 }
 
 @article{sharma2017nonvolatile,
@@ -90,7 +90,7 @@
   pages={262903},
   year={2020},
   publisher={AIP Publishing LLC},
-  doi:{https://doi.org/10.1063/5.0009185}
+  doi={https://doi.org/10.1063/5.0009185}
 }
 
 @article{smaabraaten2020domain,
@@ -102,7 +102,7 @@
   pages={033159},
   year={2020},
   publisher={APS},
-  doi:{https://doi.org/10.1103/PhysRevResearch.2.033159}
+  doi={https://doi.org/10.1103/PhysRevResearch.2.033159}
 }
 
 @article{smaabraaten2018charged,
@@ -114,7 +114,7 @@
   pages={114405},
   year={2018},
   publisher={APS},
-  doi:{https://doi.org/10.1103/PhysRevMaterials.2.114405}
+  doi={https://doi.org/10.1103/PhysRevMaterials.2.114405}
 }
 
 @article{ederer2005effect,
@@ -126,7 +126,7 @@
   pages={257601},
   year={2005},
   publisher={APS},
-  doi:{https://doi.org/10.1103/PhysRevLett.95.257601}
+  doi={https://doi.org/10.1103/PhysRevLett.95.257601}
 }
 
 @article{wang2013bifeo,
@@ -138,7 +138,7 @@
   pages={267601},
   year={2013},
   publisher={APS},
-  doi:{https://doi.org/10.1103/PhysRevLett.110.267601}
+  doi={https://doi.org/10.1103/PhysRevLett.110.267601}
 }
 
 @article{wang2003epitaxial,
@@ -150,7 +150,7 @@
   pages={1719--1722},
   year={2003},
   publisher={American Association for the Advancement of Science},
-  doi:{10.1126/science.1080615}
+  doi={10.1126/science.1080615}
 }
 
[…]
