### 2024-08-23 #6895 @mzkhalid039
> Congrats, @mzkhalid039, we're nearly finished! I'll send editorial comments on the paper soon; in the meantime, you can start working on the author tasks listed above and send me the relevant info (version tag and DOI) once you have them.

The new version of the software is version 2.0.0. You can access it through the Zenodo archive at the following DOI: https://doi.org/10.5281/zenodo.13367853. Additionally, I have included the DOIs for relevant papers in the paper.bib file.
