### 2024-08-26 #6895 @rkurchin
@mzkhalid039, a few things:
* please fix the DOI issues indicated above (mostly just need to remove URL prefix)
* Your repository indicates that your code is MIT licensed, but the Zenodo archive shows a CC license – these need to match!
