+++ paper/paper.md
@@ -21,9 +21,9 @@ bibliography: paper.bib
 
 # Summary
 
-In ferroelectric materials, the order parameter polarization can be switched by an external electric field. Regions within ferroelectric materials with uniform polarization are called domain, and the boundaries between  domains with differently aligned polarization vectors are called domain walls (DWs). These domain walls, which are only a few nanometers wide, possess unique properties with potential technological applications. DWs show promise for nanoscale electronic circuit elements and enable innovative design concepts because they can be created, erased and moved using applied electric fields [@meier2022ferroelectric; @catalan2012domain; @meier2015functional; @bednyakov2018physics]. DWs can also replicate the functionality of key electronic components such as diodes [@whyte2015diode], transistors [@mundy2017functional], and random access memories (RAM) [@sharma2017nonvolatile] .
+In ferroelectric materials, the order parameter polarization can be switched by an external electric field. Regions within ferroelectric materials with uniform polarization are called domains, and the boundaries between  domains with differently aligned polarization vectors are called domain walls (DWs). These domain walls, which are only a few nanometers wide, possess unique properties with potential technological applications. DWs show promise for nanoscale electronic circuit elements and enable innovative design concepts because they can be created, erased and moved using applied electric fields [@meier2022ferroelectric; @catalan2012domain; @meier2015functional; @bednyakov2018physics]. DWs can also replicate the functionality of key electronic components such as diodes [@whyte2015diode], transistors [@mundy2017functional], and random access memories (RAM) [@sharma2017nonvolatile].
 
-Due to the nanoscale sizes and promising properties of DWs, there has been significant interest in studying how to control and manipulate them using atomistic simulations [@schultheiss2020intrinsic; @smaabraaten2020domain; @smaabraaten2018charged]. However, developing atomic DW structures is challenging and requires knowledge of the order parameter and DW types in ferroelectric materials. DWs can be ferroelectric, antiferroelectric, and/or ferroelastic, and they can vary depending on the allowed symmetry of the ferroelectric material. For instance, ferroelectric BiFeO$_3$ exists at room temperature as a rhombohedrally distorted perovskite with space group R3c and spontaneous polarization oriented along the [111]$_P$ axis[@ederer2005effect; @wang2003epitaxial]. The symmetry constraints of having the ferroelectric polarization in BiFeO$_3$ along <111> directions gives four types of DWs across which the polarization direction changes by 71°, 109°, or 180°[@wang2003epitaxial]. Similarly, other domain wall types have been identified in other ferroelectric materials such as BaTiO$_3$[@taherinejad2012bloch], PbTiO$_3$[@meyer2002ab] and YMnO$_3$[@smaabraaten2018charged], and in ferroelastics like CaTiO$_3$[@barone2014improper].
[…]
