+++ paper/paper.bib
@@ -253,7 +253,7 @@
 }
 
 @inproceedings{mzkhalid2019first,
-  title={A first-principles study of the Al (001)/Fe (0-11) interface},
+  title={A first-principles study of the {Al(001)/Fe(0-11)} interface},
   author={Khalid, Muhammad Zeeshan and Friis, Jesper and Ninive, Per Harald and Marthinsen, Knut and Strandlie, Are},
   booktitle={Materials Science Forum},
   volume={941},
