### 2024-08-26 #6895 @mzkhalid039
> Editorial feedback on paper:
>
> * line 9: domain -> domains
> * 17: extra space after closed paren
> * 26, 28, 30, 31: add spaces before citations
> * 108: broken paralellism, remove `hibuilder.py` and just start with "Develops"
> * 106-110: would suggest changing ends of first three list items to commas and ending only #4 with a period
> * 130: fix chemical element capitalization in reference

Thank you for your comments. I have addressed all your comments.
