### 2024-08-26 23f32f7
Update paper.md
