### 2024-08-26 #6895 @mzkhalid039
> One last very small thing that I missed before: can you fix the formatting of the author names in `paper.md`? They're formatted with no spaces, as in `M.Z.Khalid`, which leads to a weird rendering in the PDF as you can see, with a space added only after the first initial. My guess is that if you change it to e.g. `M.Z. Khalid` or `M. Z. Khalid` that this should be resolved.

I have updated the names to include full first and last names, along with the initials of the middle names (`Muhammad Z. Khalid`). I think, this looks better now.
