### 2024-08-26 #6895 @rkurchin
@mzkhalid039 I think that error is being caused by the duplicated figure tag on line 84 in `paper.md`, should be easy to fix!
