### 2024-08-26 3926b56
Update paper.md
