### 2024-08-27 f92b26f
Update paper.md
