### 2024-08-27 #6895 @kyleniemeyer
Hi @mzkhalid039, a few small fixes before publishing:
- Reference "Khalid, M. Z. (2020). Atomistic modelling of Fe-Al and 𝛼-AlFeSi intermetallic compound interfaces." is missing some details.
- Reference Sharma et al. 2017 is missing its DOI: 10.1126/sciadv.1700512
