### 2024-08-27 c02febe
Update paper.bib
