+++ paper/paper.bib
@@ -78,7 +78,8 @@
   number={6},
   pages={e1700512},
   year={2017},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/sciadv.1700512}
 }
 
 @article{schultheiss2020intrinsic,
@@ -205,7 +206,8 @@
   title={Atomistic modelling of {Fe-Al} and $\alpha$-AlFeSi intermetallic compound interfaces},
   author={Khalid, Muhammad Zeeshan},
   year={2020},
-  publisher={NTNU}
+  publisher={NTNU},
+  URL={https://ntnuopen.ntnu.no/ntnu-xmlui/handle/11250/2721346}
 }
 
 @article{khalid2020ab,
