### 2024-08-27 #6895 @kyleniemeyer
@mzkhalid039 for the thesis, can you add a bit more info? A full BibTeX entry would be:
```bibtex
@phdthesis{citekey,
  author  = "",
  title   = "",
  school  = "",
  year    = "",
  type    = "",
  address = "",
  month   = "",
}
```
so adding university, type (e.g., "Ph.D. dissertation"), location would be helpful.
