+++ paper/paper.bib
@@ -202,12 +202,14 @@
   doi={10.1016/j.scriptamat.2005.01.040}
 }
 
-@article{khalid2020atomistic,
-  title={Atomistic modelling of {Fe-Al} and $\alpha$-AlFeSi intermetallic compound interfaces},
-  author={Khalid, Muhammad Zeeshan},
-  year={2020},
-  publisher={NTNU},
-  URL={https://ntnuopen.ntnu.no/ntnu-xmlui/handle/11250/2721346}
+@phdthesis{khalid2020atomistic,
+  author  = {Khalid, Muhammad Zeeshan},
+  title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
+  school  = {NTNU},
+  year    = {2020},
+  type    = {Ph.D. dissertation},
+  address = {Trondheim, Norway},
+  month   = {March},
 }
 
 @article{khalid2020ab,
