+++ paper/paper.bib
@@ -203,13 +203,13 @@
 }
 
 @phdthesis{khalid2020atomistic,
-  author  = {Khalid, Muhammad Zeeshan},
-  title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
-  school  = {NTNU},
-  year    = {2020},
-  type    = {Ph.D. dissertation},
-  address = {Trondheim, Norway},
-  month   = {March},
+  author  = "Khalid, Muhammad Zeeshan",
+  title   = "Atomistic modelling of {Fe-Al} and $\alpha$-AlFeSi intermetallic compound interfaces",
+  school  = "NTNU",
+  year    = "2020",
+  type    = "Ph.D. dissertation",
+  address = "NTNU Trondheim",
+  month   = "March",
 }
 
 @article{khalid2020ab,
