+++ paper/paper.bib
@@ -209,7 +209,7 @@
   year    = "2020",
   type    = "Ph.D. dissertation",
   address = "NTNU Trondheim",
-  month   = "March",
+  month   = "March"
 }
 
 @article{khalid2020ab,
