### 2024-08-27 8c6bd94
Update paper.bib
