+++ paper/paper.bib
@@ -203,14 +203,15 @@
 }
 
 @phdthesis{khalid2020atomistic,
-  author  = "Khalid, Muhammad Zeeshan",
-  title   = "Atomistic modelling of {Fe-Al} and $\alpha$-AlFeSi intermetallic compound interfaces",
-  school  = "NTNU",
-  year    = "2020",
-  type    = "Ph.D. dissertation",
-  address = "NTNU Trondheim",
-  month   = "March"
+  author  = {Khalid, Muhammad Zeeshan},
+  title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
+  school  = {Norwegian University of Science and Technology (NTNU)},
+  year    = {2020},
+  type    = {Ph.D. dissertation},
+  address = {Trondheim, Norway},
+  month   = {March},
 }
+
 
 @article{khalid2020ab,
   title={Ab-initio study of atomic structure and mechanical behaviour of {Al/Fe} intermetallic interfaces},
