### 2024-08-27 9ee5b2e
Update paper.bib
