+++ paper/paper.bib
@@ -207,9 +207,8 @@
   title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
   school  = {Norwegian University of Science and Technology (NTNU)},
   year    = {2020},
-  type    = {Ph.D. dissertation},
-  address = {Trondheim, Norway},
-  month   = {March},
+  type    = {Doctoral thesis},
+  note    = {Doctoral thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
 }
 
 
