### 2024-08-27 46cdbd4
Update paper.bib
