### 2024-08-27 6f563e1
Update paper.bib
