+++ paper/paper.bib
@@ -205,9 +205,8 @@
 @phdthesis{khalid2020atomistic,
   author  = {Khalid, Muhammad Zeeshan},
   title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
-  school  = {Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
+  school  = { thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
   year    = {2020},
-  type    = {Doctoral thesis}
   note    = {Doctoral thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
 }
 
