+++ paper/paper.bib
@@ -205,7 +205,7 @@
 @phdthesis{khalid2020atomistic,
   author  = {Khalid, Muhammad Zeeshan},
   title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
-  school  = { thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
+  school  = {Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
   year    = {2020},
   note    = {Doctoral thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
 }
