### 2024-08-27 6c27439
Update paper.bib
