+++ paper/paper.bib
@@ -207,6 +207,7 @@
   title   = {Atomistic Modelling of {Fe-Al} and $\alpha$-AlFeSi Intermetallic Compound Interfaces},
   school  = {Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
   year    = {2020},
+  type    = {Doctoral thesis},
   note    = {Doctoral thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
 }
 
