+++ paper/paper.bib
@@ -208,7 +208,6 @@
   school  = {Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
   year    = {2020},
   type    = {Doctoral thesis},
-  note    = {Doctoral thesis, Norwegian University of Science and Technology (NTNU), Trondheim, Norway, March 2020},
 }
 
 
