### 2024-08-27 b8c3672
Update paper.bib
