### 2024-08-27 #6895 @mzkhalid039
@kyleniemeyer for some reason, the compiled version was not displaying the full information about the thesis. However, I have resolved this issue by including the necessary details after the school's name.
